petshop-backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.11
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1
python-dotenv==1.0.0
Flask-Caching==2.1.0
redis==5.0.1
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
python-dateutil==2.8.2